refactor(pipeline): extract retry functions to retry.go [#466]
Move runWithRetry, classifyError, and backoff from engine.go to retry.go.
math import stays in engine.go (used by runPhase calibration event).
