+++ internal/pipeline/engine.go
@@ -883,115 +883,6 @@ func (e *Engine) runPhase(ctx context.Context, phase PhaseConfig) error {
 	return e.gatePhase(phase)
 }
 
-// runWithRetry runs the phase with per-category retry limits.
-func (e *Engine) runWithRetry(ctx context.Context, phase PhaseConfig, opts runner.RunOpts) (*runner.RunResult, error) {
-	remaining := map[string]int{
-		"transient": phase.Retry.Transient,
-		"parse":     phase.Retry.Parse,
-		"semantic":  phase.Retry.Semantic,
-	}
-
-	attempt := 0
-	for {
-		if err := e.apiSem.Acquire(ctx); err != nil {
-			return nil, fmt.Errorf("engine: phase %s semaphore acquire: %w", phase.Name, err)
-		}
-		result, err := e.runner.Run(ctx, opts)
-		e.apiSem.Release()
-
-		if err == nil {
-			return result, nil
-		}
-
-		category := classifyError(err)
-
-		left, tracked := remaining[category]
-		if !tracked || left <= 0 {
-			return nil, &RetriesExhaustedError{
-				Phase:    phase.Name,
-				Category: category,
-				Attempts: attempt + 1,
-				Err:      err,
-			}
-		}
-		remaining[category]--
-
-		switch category {
-		case "transient":
-			delay := backoff(attempt, e.config.JitterFunc)
-			e.config.SleepFunc(delay)
-			retryData := map[string]any{"category": category, "attempt": attempt + 1, "delay": delay.String()}
-			if suggestion := transientSuggestion(err); suggestion != "" {
-				retryData["suggestion"] = suggestion
-			}
-			e.emit(Event{
-				Phase: phase.Name,
-				Kind:  EventPhaseRetrying,
-				Data:  retryData,
-			})
-
-		case "parse":
-			var pe *runner.ParseError
-			if errors.As(err, &pe) {
-				opts.UserPrompt = opts.UserPrompt + "\n\n[RETRY] Previous attempt failed with parse error: " + pe.Error() + "\nPlease fix the output format."
-			}
-			e.emit(Event{
-				Phase: phase.Name,
-				Kind:  EventPhaseRetrying,
-				Data:  map[string]any{"category": category, "attempt": attempt + 1},
-			})
-
-		case "semantic":
-			var se *runner.SemanticError
-			if errors.As(err, &se) {
-				opts.UserPrompt = opts.UserPrompt + "\n\n[RETRY] Previous attempt returned a semantic error: " + se.Message + "\nPlease address this issue."
-			}
-			e.emit(Event{
-				Phase: phase.Name,
-				Kind:  EventPhaseRetrying,
-				Data:  map[string]any{"category": category, "attempt": attempt + 1},
-			})
-		}
-
-		_ = e.state.WriteLog(phase.Name, fmt.Sprintf("retry_%d", attempt+1),
-			[]byte(fmt.Sprintf("category=%s err=%s", category, err)))
-
-		attempt++
-	}
-}
-
-// classifyError maps an error to a retry category using agent-agnostic
-// runner error types. Backend runners (Claude, sandbox) are responsible
-// for wrapping their specific errors into runner.* types before returning.
-func classifyError(err error) string {
-	if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-		return "context"
-	}
-	var te *runner.TransientError
-	if errors.As(err, &te) {
-		return "transient"
-	}
-	var pe *runner.ParseError
-	if errors.As(err, &pe) {
-		return "parse"
-	}
-	var se *runner.SemanticError
-	if errors.As(err, &se) {
-		return "semantic"
-	}
-	return "unknown"
-}
-
-// backoff returns an exponential backoff duration capped at 30s, plus jitter.
-func backoff(attempt int, jitterFunc func(time.Duration) time.Duration) time.Duration {
-	base := 2 * time.Second
-	exp := time.Duration(math.Pow(2, float64(attempt))) * base
-	if exp > 30*time.Second {
-		exp = 30 * time.Second
-	}
-	return exp + jitterFunc(time.Second)
-}
-
 // sleepWithContext runs SleepFunc(d) but returns early if ctx is cancelled.
 // It delegates to e.config.SleepFunc in a goroutine so that test no-op sleeps
 // complete instantly while production sleeps remain cancellable.
